requirements.txt
Django==1.8
django-material==0.7.0
PyMySQL==0.7.2
